Baseline Models/requirements.txt
# Core ML Libraries
torch>=2.2.0
torchvision>=0.17.0
scikit-learn>=1.4.0

# Data Handling
numpy>=1.26.0
pandas>=2.1.0

# Image/Video Processing
opencv-python>=4.8.0
Pillow>=10.0.0
ffmpeg-python>=0.2.0
yt-dlp>=2023.11.16

# Visualization / Plotting
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
tqdm>=4.66.0
